requirements.txt
streamlit
pandas
gspread
google-auth
google-genai
plotly
fpdf
openpyxl